- Created the SearchHelpPointEarnedAchievementGateway.kt.

+++ home-helper-api/src/main/kotlin/home/helper/api/persistence/database/mysql/help_point/search/SearchHelpPointResult.kt
@@ -10,6 +10,6 @@ data class SearchHelpPointResult(
     var id: Long? = null,
     var userId: Long? = null,
     var earnedDate: LocalDate? = null,
-    var earnedPoint: Int? = null,
+    var earnedPoint: Long? = null,
     var detailList: List<HelpItemResult>? = null,
 )
